requirements.txt
ply==3.11
pytest==8.4.2
pywebview==5.2
